Add .strip to API Key loading call
Testers received error that the call could not complete, obviously pointing to the Bearer token on the call to the OpenAI API. We noticed that even if the file was obviously a single line, and was created with e.g. echo [apiKey] > openaiapikey.txt  the error would occur. Somehow a newline is being appended somewhere and being an annoyance, and this fixes that proactively.